Fix errors POSTMAN
