--- src/main/java/ru/yandex/practicum/model/FilmRequest.java
@@ -1,34 +1,0 @@
-package ru.yandex.practicum.model;
-
-import jakarta.validation.constraints.*;
-import lombok.Data;
-
-import java.time.LocalDate;
-import java.util.HashSet;
-import java.util.Set;
-
-@Data
-public class FilmRequest {
-    @NotBlank
-    private String name;
-
-    @Size(max = 200)
-    private String description;
-
-    @NotNull
-    @PastOrPresent
-    private LocalDate releaseDate;
-
-    @Positive
-    private int duration;
-
-    @NotNull
-    private MpaDto mpa;
-
-    private Set<GenreDto> genres = new HashSet<>();
-
-    @AssertTrue(message = "Release date must be after 1895-12-28")
-    public boolean isReleaseDateValid() {
-        return releaseDate.isAfter(LocalDate.of(1895, 12, 28));
-    }
-}
